requirements.txt
# Core dependencies
flask>=3.0.0
flask-cors>=4.0.0
aiohttp>=3.10.3

# Browser automation
playwright>=1.45.0

# LLM integration
langchain-ollama>=0.1.0

# Vision and OCR
opencv-python>=4.10.0.84
pillow>=11.0.0
pytesseract==0.3.10
numpy>=2.0.0

# Utilities
websockets>=12.0
requests>=2.31.0
python-dotenv>=1.0.0

# Data processing

# Testing
pytest>=7.0.0
